+++ src/app.rs
@@ -7,6 +7,9 @@ use ratatui::{
     },
 };
 
+#[cfg(feature = "dev-console")]
+use crate::logging;
+
 use crate::rendering;
 use crate::stage_handlers::{Stage, StageHandler};
 use crate::{game::Game, stage_handlers::ReadyHandler};
@@ -25,7 +28,6 @@ impl App {
     pub fn run(mut self, mut terminal: DefaultTerminal) -> anyhow::Result<()> {
         log::info!("Main loop is starting...");
 
-        crate::dev_gray!("### {:?}", self.stage);
         while self.is_running {
             terminal
                 .draw(|frame| {
@@ -52,7 +54,6 @@ impl App {
 
     fn tick(&mut self) {
         if let Some(next_stage) = self.stage.update(&mut self.game) {
-            crate::dev_gray!("/// {next_stage:?}");
             self.stage = next_stage;
         }
     }
@@ -67,7 +68,6 @@ impl App {
             Event::Mouse(mouse_event)
                 if matches!(mouse_event.kind, crossterm::event::MouseEventKind::ScrollUp | crossterm::event::MouseEventKind::ScrollDown) =>
             {
-                use crate::logging;
                 logging::dev_console::handle_mouse_scroll_event(*mouse_event);
             }
             _ => (),
@@ -84,7 +84,6 @@ impl App {
         // "dev-console" keys
         #[cfg(feature = "dev-console")]
         {
-            use crate::logging;
             if logging::dev_console::handle_key_pressed_event(key_event) {
                 return;
             }
@@ -92,7 +91,6 @@ impl App {
 
         // Stages keys
         if let Some(next_stage) = self.stage.handle_key_pressed_event(&mut self.game, *key_event) {
-            crate::dev_gray!("--> {next_stage:?}");
             self.stage = next_stage;
         }
     }