+++ src/rendering/paused_ui.rs
@@ -6,11 +6,8 @@ use super::{LegendItem, compile_legend, draw_keys_legend};
 
 #[rustfmt::skip]
 const LEGEND_ITEMS: &[LegendItem] = &[
-    LegendItem { key: "Arrows",   action: "Move/Rotate" },
-    LegendItem { key: "0 (Zero)", action: "Accelerate" },
-    LegendItem { key: "Space",    action: "Drop" },
-    LegendItem { key: "Esc",      action: "Pause" },
-    LegendItem { key: "Q",        action: "Quit" },
+    LegendItem { key: "Enter", action: "Resume" },
+    LegendItem { key: "Q",     action: "Quit" },
 ];
 
 static LEGEND: LazyLock<(Text<'_>, Text<'_>)> = LazyLock::new(|| compile_legend(LEGEND_ITEMS));