+++ src/rendering/ready_ui.rs
@@ -1,9 +1,8 @@
 use std::sync::LazyLock;
 
-use ratatui::{Frame, text::Text};
+use ratatui::{Frame, layout::Rect, text::Text};
 
-use super::{LayoutAreas, LegendItem, compile_legend, draw_board, draw_keys_legend, draw_next_column, draw_stats};
-use crate::game::Game;
+use super::{LegendItem, compile_legend, draw_keys_legend};
 
 #[rustfmt::skip]
 const LEGEND_ITEMS: &[LegendItem] = &[
@@ -14,18 +13,6 @@ const LEGEND_ITEMS: &[LegendItem] = &[
 
 static LEGEND: LazyLock<(Text<'_>, Text<'_>)> = LazyLock::new(|| compile_legend(LEGEND_ITEMS));
 
-pub(super) fn render(frame: &mut Frame, game: &Game, layout_areas: &LayoutAreas) {
-    // TODO refactor: encapsulate in mod.rs?
-    draw_next_column(frame, layout_areas.next_column, game);
-    draw_board(frame, layout_areas.board, game);
-    draw_stats(frame, layout_areas.stats, game);
-
-    draw_keys_legend(frame, layout_areas.footer, &LEGEND);
-
-    // TODO refactor: call in mod.rs?
-    #[cfg(feature = "dev-console")]
-    {
-        use crate::logging;
-        logging::dev_console::draw(frame, layout_areas.dev_console);
-    }
+pub(super) fn render(frame: &mut Frame, footer_area: Rect) {
+    draw_keys_legend(frame, footer_area, &LEGEND);
 }